+++ src/main.py
@@ -1,0 +1,8 @@
+from .BetterScratch import BetterScratch
+
+def main():
+    app=BetterScratch()
+    app.mainloop()
+
+if __name__=="__main__":
+    main()
